livedata final

+++ app/src/main/java/com/hishamabifarah/androidx_kotlin/mvvmMoviesApp/mvvmMovieData/model/MovieDetails.kt
@@ -18,7 +18,7 @@ data class MovieDetails(
     @SerializedName("poster_path") val posterPath: String,
     @SerializedName("production_companies") val productionCompanies: List<ProductionCompany>,
     @SerializedName("release_date") val releaseDate: String,
-    @SerializedName("revenue") val revenue: Int,
+    @SerializedName("revenue") val revenue: Long,
     @SerializedName("runtime") val runtime: Int,
     @SerializedName("status") val status: String,
     @SerializedName("tagline") val tagline: String,
